Add the ability for `git-duet` to optionally set user.name and user.email
When GIT_DUET_SET_GIT_USER_CONFIG is set to 1, also set `user.name` and
`user.email`. This allows commands like `git rebase` which require
an author to be set in `user.name` and `user.email` to still be able to
function.

Default to off for now, but this may change in the future.

Also update one test for error message propogation to use compare to
running `git config` since differing versions of `git` return different
error messages..