Make all commands support GIT_DUET_GLOBAL
This does the following:
* Make `git solo` and `git duet` with no initials respect `-g` and
  `GIT_DUET_GLOBAL` when displaying configuration
* Make `git duet-(merge|revert|commit)` respect `GIT_DUET_GLOBAL`. If
  `GIT_DUET_GLOBAL` is set, they will ignore any repo level git-duet
  configuration.

Instead of unsetting keys for the committers when running `git solo`,
set them to empty string to avoid falling through to the global git
config when reading configuration (i.e. if the author was set locally,
but no committer it would fall back to the global committer which is
probably not desirable).

Fixes #14